Debug

+++ src/main/java/org/example/cache/storage/DiskStorage.java
@@ -5,9 +5,15 @@ import org.example.cache.exceptions.FileAccessException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.io.*;
+import java.io.BufferedInputStream;
+import java.io.BufferedOutputStream;
+import java.io.File;
+import java.io.FileInputStream;
+import java.io.FileOutputStream;
+import java.io.IOException;
+import java.io.ObjectInputStream;
+import java.io.ObjectOutputStream;
 import java.util.Arrays;
-import java.util.Comparator;
 import java.util.List;
 
 public class DiskStorage<T, V> implements StorageStrategy<T, V> {
@@ -24,7 +30,7 @@ public class DiskStorage<T, V> implements StorageStrategy<T, V> {
 
     @Override
     public void put(T key, V value) throws FileAccessException {
-        if (size == cacheFiles.listFiles().length) pruning();
+        pruning();
         try (ObjectOutputStream objectOutputStream = new ObjectOutputStream(new BufferedOutputStream(new FileOutputStream(DIRECTORY + key + FILETYPE)))) {
             objectOutputStream.writeObject(value);
             LOGGER.info("the file was successfully written. Key: " + key + "Value: " + value);
@@ -36,35 +42,39 @@ public class DiskStorage<T, V> implements StorageStrategy<T, V> {
 
     @Override
     public V get(T key) throws FileAccessException {
-        V value;
         try (ObjectInputStream objectInputStream = new ObjectInputStream(new BufferedInputStream(new FileInputStream(DIRECTORY + key + FILETYPE)))) {
-            value = (V) objectInputStream.readObject();
+            V value = (V) objectInputStream.readObject();
             File file = new File(DIRECTORY + key + FILETYPE);
             if (file.exists()) {
                 file.setLastModified(System.currentTimeMillis());
             }
             LOGGER.info("the file was successfully received");
+            return value;
         } catch (IOException | ClassNotFoundException e) {
             LOGGER.error("Error in class DiskStorage, method get. Error is related to get object from file or file not exists");
             throw new FileAccessException("Could not read object from " + key + FILETYPE + " in " + DIRECTORY);
         }
-        return value;
     }
 
     private void pruning() {
-        List<File> listWithFiles = Arrays.asList(cacheFiles.listFiles().clone());
-        Comparator<File> fileComparator = Comparator.comparing(File::lastModified);
-        listWithFiles.sort(fileComparator);
-        File minLastModified = listWithFiles.get(0);
-        minLastModified.delete();
-        LOGGER.info("the file was deleted following the LRU strategy");
+        if (cacheFiles.exists() && cacheFiles.listFiles() != null && size <= cacheFiles.listFiles().length) {
+            List<File> listWithFiles = Arrays.asList(cacheFiles.listFiles().clone());
+            listWithFiles.sort(((o1, o2) -> Long.compare(o2.lastModified(), o1.lastModified())));
+            for (int i = size - 1; i < listWithFiles.size(); i++) {
+                File minLastModified = listWithFiles.get(i);
+                minLastModified.delete();
+            }
+            LOGGER.info("the file was deleted following the LRU strategy");
+        }
     }
 
     @Override
     public void clear() {
-        for (File myFile : cacheFiles.listFiles()) {
-            if (myFile.isFile()) myFile.delete();
+        if (cacheFiles.exists() && cacheFiles.listFiles() != null) {
+            for (File myFile : cacheFiles.listFiles()) {
+                if (myFile.isFile()) myFile.delete();
+            }
+            LOGGER.info("The cache has been cleared");
         }
-        LOGGER.info("The cache has been cleared");
     }
 }
